Switch from lazy to eager generation of matrices
All optical elements were previously defined with a field `mat` of type
`Option<ComplexMatrix>`. This added a layer of complexity with no
obvious benefit. A consequence of this type is that all optical elements
had to be declared as `mut`. The optical element is initialized with
`mat: None`, which means that the optical element must be mutable if the
matrix is to be stored after it is eventually generated. Requiring the
user to declare the element as `mut` when they never intend to modify it
doesn't make sense from an API standpoint, so the type was changed.