Remove unnecessary struct field 'angle'
An angle is needed to construct each optical element, but it is not
needed afterwards. There is no need to carry around the initial angle of
the element, and it would need to be updated every time the element is
rotated in-place with `rotate`, so the field was simply removed.