Overhaul traits to be more consistent
To access the Jones matrix of a type that implemented `JonesMatrix`, you
would call the `T.matrix()` method. However, the type that implemented
the `JonesVector` trait was just a bare `Vector2`, so no method call was
required to get the Jones vector. Furthermore, it was impossible to add
convenience methods for constructing Jones vectors without coming up
with a separate trait, or adding the methods to the `JonesVector` trait.
For these reasons, the `JonesVector` trait was implemented for `Beam`, which
is simply a wrapper around the `Vector2` that it contains.