Remove the IntensityTooSmall error condition
The lower limit for this error condition was 1e-12, which was chosen
arbitrarily. Furthermore, there was never really an error encountered
because an intensity calculation returned a number that was too small.
The error condition was originally added to ease property based testing,
but this is no longer a problem when reasonable values are generated to
begin with.