Reimplement tests using the Arbitrary trait
I encountered a bug in tests related to `remove_common_phase` that would
expect the wrong phase to be returned for the y-component. Complex
numbers are stored as real and imaginary parts, not magnitude and phase.
When a complex number is constructed from a magnitude and phase, the
real and imaginary parts are computed from mag*cos(phase) and
mag*sin(phase). If the magnitude is zero, then both the real and
imaginary parts are zero. Zeroing out the imaginary part of the complex
number wipes out the phase. This means that the phase you used to
construct the complex number is not the phase you get back (0) when you
ask for the phase. The tests now account for this, and also adjust the
expected phase to be in the range [-pi, +pi].