Allow OpticalElements to return the inner matrix
The OpticalElement enum is a wrapper around various types that implement
the JonesMatrix trait. This wrapper allows you to make a vector of
optical elements without needing to specify the type of the vector as
Vec<impl JonesMatrix>, since that isn't allowed. The wrapper does make
it tedious to access the actual matrix of the optical element, so this
method allows you to access that matrix in a more convenient way.