Fix a bug that prevents adding new elements
This commit fixes a bug that would silently prevent you from adding new
elements to an optical system i.e. they would be completely ignored.
The bug was located in the `with_element` method. The original code is shown below:

```
let elements = match self.elements {
    Some(elements) => {
        elements.clone().push(elem);
        elements
    },
    None => vec![elem],
};
```

When there are elements already in the system, a new element is supposed
to be pushed onto the vector containing the elements. Rather than deal
with lifetime issues, the vector was intended to be cloned, the new
element was to be pushed onto the clone, and the clone was to be
returned.

The bug lies in these two lines:
```
elements.clone().push(elem)
elements
```
Here the vector of elements is cloned, but the clone is never bound to
an identifier. The new element is pushed onto the vector, which is then
immediately dropped along with the new element. Furthermore, the
original vector is returned, rather than the clone.

The new code make a very simple modification:
```
let mut new_elements = elements.clone();
new_elements.push(elem);
new_elements
```
This could probably be done in a single line without binding the cloned
vector to a name, but binding it to a name makes it easier to look at in
a debugger.